Add view orders, isAdmin authenticating function

+++ src/app/api/categories/route.ts
@@ -1,6 +1,7 @@
 import Category from "@/app/models/Category";
 import MenuItem from "@/app/models/MenuItems";
 import { NextRequest, NextResponse } from "next/server";
+import { isAdmin } from "../auth/[...nextauth]/route";
 
 export async function GET(req: NextRequest) {
     try {
@@ -13,16 +14,21 @@ export async function GET(req: NextRequest) {
 
 export async function POST(req: NextRequest) {
     try {
-        const d = await req.json();
+
+        if(await isAdmin()){
+            const d = await req.json();
         
-        const newCategory = new Category({
-            name: d.categoryName
-        });
+            const newCategory = new Category({
+                name: d.categoryName
+            });
 
-        const savedCategory = await newCategory.save();
+            const savedCategory = await newCategory.save();
 
-        return NextResponse.json(savedCategory);
-        // return NextResponse.json(true);
+            return NextResponse.json(savedCategory);
+        } else {
+            return NextResponse.json({error: 'Cannot see the requested page.'}, {status: 401});
+        }
+
     }catch(err: any){
         return NextResponse.json({error: err.message}, {status: 500});
     }
@@ -30,15 +36,17 @@ export async function POST(req: NextRequest) {
 
 export async function PUT(req: NextRequest) {
     try {
-        const data = await req.json();
 
-        if('categoryName' in data && '_id' in data){
-            
-            const res = await Category.findByIdAndUpdate(data._id, {name: data.categoryName}, {new: true});            
-
-            return NextResponse.json(res);
+        if(await isAdmin()){
+            const data = await req.json();    
+            if('categoryName' in data && '_id' in data){
+                const res = await Category.findByIdAndUpdate(data._id, {name: data.categoryName}, {new: true});            
+                return NextResponse.json(res);
+            } else {
+                throw new Error('Something went wrong!');
+            }
         } else {
-            throw new Error('Something went wrong!');
+            return NextResponse.json({error: 'Cannot see the requested page.'}, {status: 401});
         }
 
     } catch (error: any) {
@@ -48,22 +56,24 @@ export async function PUT(req: NextRequest) {
 
 export async function DELETE(req: NextRequest) {
     try {
-        const data = await req.json();
-
-        if('_id' in  data){
-            const cat = await Category.findById(data._id);
-            const menuItems = await MenuItem.find({category: cat.name});
-            
-            if(menuItems.length > 0){
-                throw new Error(`${menuItems.length} Menu Items exists with this Category.`);
+        if(await isAdmin()){
+            const data = await req.json();
+            if('_id' in  data){
+                const cat = await Category.findById(data._id);
+                const menuItems = await MenuItem.find({category: cat.name});
+                
+                if(menuItems.length > 0){
+                    throw new Error(`${menuItems.length} Menu Items exists with this Category.`);
+                }
+            }   
+            if('_id' in data){
+                const res = await Category.findByIdAndDelete(data._id);            
+                return NextResponse.json(res);
+            } else {
+                throw new Error('Something went wrong!');
             }
-        }
-
-        if('_id' in data){
-            const res = await Category.findByIdAndDelete(data._id);            
-            return NextResponse.json(res);
         } else {
-            throw new Error('Something went wrong!');
+            return NextResponse.json({error: 'Cannot see the requested page.'}, {status: 401});
         }
     } catch (error: any) {
         return NextResponse.json({error: error.message}, {status: 500});
